+++ client/src/components/AgentComponents/AgentPatientSearchBar.tsx/AgentPatientSearchBar.type.tsx
@@ -1,4 +1,3 @@
 export default interface AgentPatientSearchBarProps {
   handlePatientSelected: (patientId: number) => void;
-  restriction?: boolean;
 }
