+++ client/src/pages/AgentHome/AgentHome.tsx
@@ -115,7 +115,6 @@ export default function AgentHome() {
         handlePatientSelected={function (patientId: number): void {
           console.info(`Patient ID sélectionné  : ${patientId}`);
         }}
-        restriction={true}
       />
       <button
         className="mt-4 rounded bg-gray-500 px-4 py-2 text-white"
